tflite converter optimization datatype changed to float32

+++ ML/tflite_converter.py
@@ -3,7 +3,7 @@ import tensorflow as tf
 saved_model_dir = r'C:\Users\Ankan\Desktop\Github\ML-with-Flutter\ML\best_model'
 converter = tf.lite.TFLiteConverter.from_saved_model(saved_model_dir)
 converter.optimizations = [tf.lite.Optimize.DEFAULT]
-converter.target_spec.supported_types = [tf.float16]
+converter.target_spec.supported_types = [tf.float32]
 tflite_model = converter.convert()
 
 with open('converted_model_tflite', 'wb') as f:
